requirements.txt
Django~=5.0
psycopg2-binary~=2.9
djangorestframework~=3.15
drf-spectacular~=0.27
django-filter~=24.2